### 2021-04-30 #3231 @whedon
[PRE REVIEW]: funsies: A minimalist, distributed and dynamic workflow engine
**Submitting author:** @clavigne (<a href="http://orcid.org/0000-0003-2778-1866">Cyrille Lavigne</a>)
**Repository:** <a href="https://github.com/aspuru-guzik-group/funsies" target ="_blank">https://github.com/aspuru-guzik-group/funsies</a>
**Version:** v0.7.0
