### 2021-05-08 #3231 @clavigne
sorry for the delay, Ive had a very busy week.

I'd recommend [Johannes Koester](https://github.com/johanneskoester) or someone from his group, [Paolo Di Tommaso](https://github.com/pditommaso) and [Hilary J. Oliver](https://github.com/hjoliver). They are all working on similar codes (inspirations of sort).

Thank you!
