### 2021-05-10 de0289b
made type inference warning more useful
