### 2021-05-11 26a9cd9
graph now properly include subdag children
