### 2021-05-11 #3274 @whedon
[REVIEW]: funsies: A minimalist, distributed and dynamic workflow engine
**Submitting author:** @clavigne (<a href="http://orcid.org/0000-0003-2778-1866">Cyrille Lavigne</a>)
**Repository:** <a href="https://github.com/aspuru-guzik-group/funsies" target ="_blank">https://github.com/aspuru-guzik-group/funsies</a>
**Version:** 0.8.1
