### 2021-05-13 a692b03
fixed bug for waiting on links
