+++ src/funsies/_graph.py
@@ -42,8 +42,11 @@ class ArtefactStatus(IntEnum):
     linked = 4
 
 
-def get_status(db: Redis[bytes], address: hash_t) -> ArtefactStatus:
+def get_status(db: Redis[bytes], address: hash_t, resolve_links=False) -> ArtefactStatus:
     """Get the status of an artefact."""
+    if resolve_links:
+        address = resolve_link(db, address)
+
     val = db.get(join(ARTEFACTS, address, "status"))
     if val is None:
         return ArtefactStatus.not_found
