### 2021-05-21 #3274 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1063/5.0006002 is OK
> - 10.1063/5.0004835 is OK
> - 10.1063/5.0007045 is OK
> - 10.1063/5.0004997 is OK
> - 10.1063/5.0005188 is OK
> - 10.1063/1.5144261 is OK
> - 10.1016/j.matt.2021.02.017 is OK
> - 10.1145/1330017.1330019 is OK
> - 10.1093/bioinformatics/bts480 is OK
> - 10.12688/f1000research.29032.2 is OK
> - 10.1145/2807591.2807623 is OK
> - 10.1007/10968987_3 is OK
> - 10.1007/978-3-319-63688-7_19 is OK
>
> MISSING DOIs
>
> - 10.26434/chemrxiv.13008500.v1 may be a valid DOI for title: Automatic discovery of chemical reactions using imposed activation
> - 10.21276/ijre.2018.5.5.4 may be a valid DOI for title: MapReduce: Simplified Data Processing on Large Clusters
>
> INVALID DOIs
>
> - None
> ```

@clavigne please add the missing DOIs to the paper and recompile the paper.
