### 2021-05-25 #3274 @clavigne
>
>
> I added the DOI for the Chemrxiv pre-print. The other paper, (which is available here https://research.google/pubs/pub62/ ) does not have a DOI, as far as I can tell.
>
> The ACM which compiles the USENIX proceedings doesn't have the OSDI '04 one https://dl.acm.org/conference/osdi/proceedings .

@diehlpk as above, there are no DOI for this reference, as far as I know.
