### 2021-05-25 #3274 @diehlpk
@clavigne

>
> MISSING DOIs
>
> - 10.21276/ijre.2018.5.5.4 may be a valid DOI for title: MapReduce: Simplified Data Processing on Large Clusters
