### 2021-05-26 #3274 @gflofst
added issue for making community guidelines more apparent. Continuing other testing.
