+++ README.md
@@ -156,3 +156,8 @@ other worfklow engine can be found
 ## License
 
 funsies is provided under the MIT license.
+
+## Contributing
+
+All contributions are welcome! Consult [the CONTRIBUTING](./CONTRIBUTING.md)
+file for help. Please file issues for bugs and documentation issues.
