+++ CONTRIBUTING.md
@@ -7,16 +7,18 @@ If you encounter an issue, please fill a report [using the github Issues
 page.](https://github.com/aspuru-guzik-group/funsies/issues)
 
 If you have a fix, use the [pull request functionality on
-Github.](https://github.com/aspuru-guzik-group/funsies/pulls). Make sure your
-commits are named properly.
+Github.](https://github.com/aspuru-guzik-group/funsies/pulls) Make sure your
+commits are given concise, explanatory names and limit your file edits to the
+minimal relevant parts. Always run formatters before comitting (as described
+below).
 
-## Tests and lints
+## Instructions for contributors
 
 This repo has a CI workflow that will run the test suite, mypy, linting etc.,
-and all tests and lints should pass. If you want to test it before pushing to
-the CI, all of this is [automated using
-Nox.](https://nox.thea.codes/en/stable/) To install nox and the development
-packages in a python env, use
+and all tests and lints should pass. If you want to run CI locally before
+pushing, all of it is [automated using Nox for local
+development.](https://nox.thea.codes/en/stable/) To install nox in a python
+env, use
 
 ```bash
 pip install nox
@@ -38,24 +40,23 @@ nox -rs mypy
 Tests can be run using `nox -rs tests` which will run tests for three python
 versions (if they are installed). Some of the tests are rather slow and so are
 skipped by default, you can run the whole set using `nox -rs tests -- --cov`.
-You'll need a redis server installed and available on `$PATH`.
+You'll need a redis server installed and available on `$PATH`. 
 
-## Code conventions and best practices
+Note that nox will take care of installing packages from PyPI for each of the
+above steps, so you shouldn't need to do anything besides installing nox
+itself to get a working dev environment.
 
 funsies is formatted using [black](https://github.com/psf/black) and
 [isort](https://pypi.org/project/isort/), which are run automatically using
 nox, as described above.
 
-Importantly, funsies is (as much as possible in python) a statically typed
-code, and every function boundaries needs annotations. In addition, because
-funsies is used to simulate large, multi-node workflows, raising exceptions
-should be avoided as much as possible.
+Importantly, **funsies is a statically typed program,** and every function
+boundaries needs annotations. 
 
-And note that (critically) the code tries to maintain (as much as possible) a
-functionally pure paradigm for operations that are not encoded in the
-database. The lower-level architecture of this system is coded in
-[_graph.py](https://github.com/aspuru-guzik-group/funsies/blob/master/src/funsies/_graph.py)
-and should not be much of a concern if you don't touch this code.
+Note also that the code tries to maintain (as much as possible) a functionally
+pure paradigm for operations encoded in the database. This means that adding
+new functionality can be somewhat tricky. If you find yourself meddling with
+the internals described below, please talk to me (@clavigne) first!
 
 
 ## Internals: how to read this code
@@ -64,8 +65,9 @@ funsies is not a particularly complex codebase (yet) and should be fairly
 readable. Here is a basic summary of the internal architecture.
 
 The low-level workflow generation and encoding can be found in
-[_graph.py](https://github.com/aspuru-guzik-group/funsies/blob/master/src/funsies/_graph.py)and
-[_funsies.py](https://github.com/aspuru-guzik-group/funsies/blob/master/src/funsies/_funsies.py).
+[_graph.py](https://github.com/aspuru-guzik-group/funsies/blob/master/src/funsies/_graph.py)
+ and
+[_funsies.py](https://github.com/aspuru-guzik-group/funsies/blob/master/src/funsies/_funsies.py). 
 The former contains hash calculation, getters and setters for data and
 dependencies, while the latter contains data structure for encoding
 operations. Data is serialized in
