### 2021-05-28 #3274 @gflofst
Everything works for me. The last bit of testing on Windows getting an up to date Redis working is going to take time. As it is the last bit, and everything else is good, I am ok signing off on it for now and I'll raise an issue later if it isn't working. Windows is a marginal platform for this field, but it is what I have for my desktop machine for testing.
