### 2021-06-09 2248c71
Squashed commit of the following:
commit bfa8ae918fa4b06080860c04df5fdd50cf252413
Merge: 1daf4e4b 4bda10b9
Author: Cyrille Lavigne <clavigne@chem.utoronto.ca>
Date:   Wed May 26 12:10:19 2021 -0400

    Merge pull request #2 from aspuru-guzik-group/feat_contributing

commit 4bda10b990a48fb06f1a4621f984bef862070a31
Author: Cyrille Lavigne <clavigne@chem.utoronto.ca>
Date:   Wed May 26 12:05:43 2021 -0400

    fixed typos and grammar

commit ba35b9e92004b54190a59a5eab35829999761b46
Author: Cyrille Lavigne <clavigne@chem.utoronto.ca>
Date:   Wed May 26 11:56:14 2021 -0400

    fixed README

commi
[…]
