+++ paper/paper.bib
@@ -29,6 +29,7 @@
 
 @article{lavigne_automatic_2020,
 	title = {Automatic discovery of chemical reactions using imposed activation},
+    doi = {10.26434/chemrxiv.13008500.v2},
 	url = {/articles/preprint/Automatic_discovery_of_chemical_reactions_using_imposed_activation/13008500/1},
 	abstract = {Computational power and quantum chemical methods have improved immensely since computers were first applied to the study of reactivity, but the de novo prediction of chemical reactions has remained challenging. We show that complex reactions can be efficiently and autonomously predicted using chemical activation imposed by simple geometrical constraints. Our approach is demonstrated on realistic and challenging chemistry, such as a triple cyclization cascade involved in the total synthesis of a natural product and several oxidative addition reactions of complex drug-like molecules. Notably and in contrast with traditional hand-guided computational chemistry calculations, our method requires minimal human involvement and no prior knowledge of products or mechanisms. Imposed activation can be a transformational tool to screen for chemical reactivity and mechanisms as well as to study byproduct formation and decomposition.},
 	author = {Lavigne, Cyrille and dos Passos Gomes, Gabriel and Pollice, Robert and Aspuru-Guzik, Alan},
@@ -317,5 +318,6 @@
 	pages = {137--150},
 	booktitle = {{OSDI}'04: Sixth Symposium on Operating System Design and Implementation},
 	author = {Dean, Jeffrey and Ghemawat, Sanjay},
+    doi = {},
 	date = {2004},
 }
