+++ src/funsies/_graph.py
@@ -5,25 +5,14 @@ from __future__ import annotations
 from dataclasses import dataclass
 from enum import IntEnum
 import hashlib
-import io
-import traceback
-from typing import (
-    Any,
-    Generic,
-    Mapping,
-    Optional,
-    Type,
-    TypeVar,
-    overload,
-    Union,
-)
+from typing import Any, Generic, Mapping, Optional, overload, Type, TypeVar, Union
 
 # python 3.7 imports Literal from typing_extensions
 try:
     # std
     from typing import Literal
 except ImportError:
-    from typing_extensions import Literal
+    from typing_extensions import Literal  # type:ignore
 
 # external
 from redis import Redis
@@ -31,23 +20,13 @@ from redis.client import Pipeline
 
 # module
 from . import _serdes
-from ._constants import _Data, ARTEFACTS, BLOCK_SIZE, Encoding, hash_t, join, OPERATIONS
+from ._constants import _Data, ARTEFACTS, Encoding, hash_t, join, OPERATIONS
 from ._funsies import Funsie
 from ._logging import logger
 from ._short_hash import hash_save
+from ._storage import get_location, Location
 from .config import Options
 from .errors import Error, ErrorKind, match, Result
-
-# Max redis value size in bytes
-MIB = 1024 * 1024
-MAX_VALUE_SIZE = 512 * MIB
-
-
-@dataclass
-class StorageUnit:
-
-    artefact: hash_t
-    key: str
 
 
 # --------------------------------------------------------------------------------
@@ -201,18 +180,12 @@ def resolve_link(db: Redis[bytes], address: hash_t) -> hash_t:
         return resolve_link(db, out)
 
 
-def _set_block_size(n: int) -> None:
-    """Change block size."""
-    global BLOCK_SIZE
-    BLOCK_SIZE = n
-
-
 def __get_storage_read(
     store: Redis[bytes],
     address: hash_t,
     carry_error: Optional[hash_t] = None,
     do_resolve_link: bool = True,
-) -> Result[StorageUnit]:
+) -> Result[Location]:
     """Perform all the prior step before actually retrieving data."""
     if do_resolve_link:
         address = resolve_link(store, address)
@@ -236,22 +209,14 @@ def __get_storage_read(
             source=carry_error,
         )
     else:
-        key = join(ARTEFACTS, address, "data")
-        if not store.exists(key):
-            return Error(
-                kind=ErrorKind.Mismatch,
-                details="expected data was not found",
-                source=carry_error,
-            )
-        else:
-            return StorageUnit(key=key, artefact=address)
+        return get_location(store, address)
 
 
 def __get_storage_write(
     store: Redis[bytes],
     address: hash_t,
     do_resolve_link: bool = True,
-) -> Result[StorageUnit]:
+) -> Result[Location]:
     """Perform all the prior step before actually retrieving data."""
     if do_resolve_link:
         address = resolve_link(store, address)
@@ -265,52 +230,7 @@ def __get_storage_write(
             kind=ErrorKind.UnresolvedLink,
             details=f"artefact at {address} is a link and thus we can't set data to it",
         )
-    key = join(ARTEFACTS, address, "data")
-    return StorageUnit(key=key, artefact=address)
-
-
-def _get_bytes(
-    store: Redis[bytes],
-    loc: Result[StorageUnit],
-) -> Result[bytes]:
-    """Retrieve bytes corresponding to an artefact."""
-    raw = match(
-        loc,
-        some=lambda x: b"".join(store.lrange(x.key, 0, -1)),
-        none=lambda x: x,
-    )
-    return raw
-
-
-def _set_bytes(
-    store: Redis[bytes],
-    loc: StorageUnit,
-    value: bytes,
[…]
