### 2021-07-14 #3274 @vijaysm
The installation for funsies was smooth on both OSX and Linux. But I couldn't get pip to install the redis-server on OSX. Had to resort to homebrew solutions. Is this expected ? Using `Python==3.8.10`.

I was able to run the example workflows just fine without issues after that initial installation hiccup.
