### 2021-07-14 #3274 @vijaysm
@clavigne  The mergesort example is interesting. However, running it gives me several error messages like below. It would be better to catch these exceptions so that the example output is cleaner.
```
 op:bf485b | runner raised!
Traceback (most recent call last):

  File "/usr/local/lib/python3.8/dist-packages/funsies/_run.py", line 186, in run_op
    out_data: out_data_t = runner(funsie, input_data)  # type:ignore
                           │      │       └ {'in0': b'[6]'}
                           │      └ Funsie(how=<FunsieHow.subdag: 2>, what='SAC|split|funsies_mergesort.<locals>.<lambda>|funsies_mergesort.<locals>.<lambda>', i...
                           └ <function run_subdag_funsie at 0x7f8dc5690280>

  File "/usr/local/lib/python3.8/dist-packages/funsies/_subdag.py", line 81, in run_subdag_funsie
    outfun = fun(decoded)
             │   └ {'in0': [6]}
             └ <function sac.<locals>.__sac at 0x7f8dc54b3670>

  File "/usr/local/lib/python3.8/dist-packages/funsies/dynamic.py", line 64, in __sac
    split_data = [constant_artefact(db, d) for d in split_fun(*args)]
                  │                 │               │          └ [[6]]
                  │                 │               └ <function split at 0x7f8dc54b38b0>
                  │                 └ Redis<ConnectionPool<Connection<host=localhost,port=6379,db=0>>>
                  └ <function constant_artefact at 0x7f8dc56d9dc0>

  File "mergesort.py", line 18, in split
    raise RuntimeError()

RuntimeError
 op:bf485b | DONE: runner raised exception.
```
If the error messages are not expected, then it should be fixed. If this is the expected behavior (?), then perhaps an explanation in documentation on why there is a `RuntimeError` being raised, and how it gets handled internally would be helpful. I can replicate this on both OSX and Linux.
