### 2021-07-19 48054cb
updated mergesort example for clarity
