### 2021-07-19 #3274 @clavigne
Hi @vijaysm , thank you for the careful review.

> The installation for funsies was smooth on both OSX and Linux. But I couldn't get pip to install the redis-server on OSX. Had to resort to homebrew solutions. Is this expected ? Using `Python==3.8.10`.
>
> I was able to run the example workflows just fine without issues after that initial installation hiccup.

I do not have access to an OSX machine to test it, so I'm not sure if this is expected. The Redis installation through pip is from this package https://github.com/pywharf/redis-server which states support for `macosx_10_15_x86_64`. At this time, Redis 4.x or above is supported by funsies. I've added this information to the Readme, as well as the description of how to install it through Homebrew.
