### 2021-07-19 f9194fa
updated mypy types to remove type:ignore statements
