+++ tests/test_errors_distributed.py
@@ -156,7 +156,7 @@ def test_job_killed(nworkers: int, sig: int) -> None:
 
 
 @pytest.mark.slow
-@pytest.mark.parametrize("nworkers", [1, 2, 8])
+@pytest.mark.parametrize("nworkers", [1, 2, 4])
 def test_data_race(nworkers: int) -> None:
     """Test a data race when execute calls are interleaved."""
     with f.ManagedFun(nworkers=nworkers):
@@ -170,8 +170,8 @@ def test_data_race(nworkers: int) -> None:
 
         f.execute(step1)
         f.execute(step2)
-        f.wait_for(step1, timeout=10.0)
-        f.wait_for(step2, timeout=10.0)
+        f.wait_for(step1, timeout=20.0)
+        f.wait_for(step2, timeout=20.0)
 
 
 @pytest.mark.slow
