### 2021-07-20 #3274 @vijaysm
@clavigne Thanks for quickly updating README.

> I do not have access to an OSX machine to test it, so I'm not sure if this is expected. The Redis installation through pip is from this package https://github.com/pywharf/redis-server which states support for macosx_10_15_x86_64. At this time, Redis 4.x or above is supported by funsies. I've added this information to the Readme, as well as the description of how to install it through Homebrew.

I didn't drill down on the reasons why pip didn't install on OSX, but showing alternative workflows to get the TPLs installed is helpful for new users. The installation information looks complete enough now.
