### 2021-07-20 #3274 @vijaysm
> The raised errors are used for control flow. They terminate the mergesort recursion dynamically without having to know in advance the required recursion depth. I'm updating the example to make this more clear, and to make it raise a more self-explanatory exception.

Got it. I assumed the runtime error was intentional, but the log messages with trace information would confuse users, and lead them to worry that either the example is broken, or their build/installation is incompatible. So displaying this in a better, more user-friendly way would be better.

You should also update README.md or appropriate documentation for funsies to describe how runtime errors are caught cleanly and redirected to the log file instead of workers quitting prematurely. This is an important feature for all real workflows and should be highlighted explicitly IMO.

Also, is there a separate documentation/example page for funsies ? I don't see one. There is a link to source docs and link to the recipes folder, but something like a user guide will add more value here as well. Bulk of what would go in a user document is already there in README.md, but separating this perhaps into its own space (github.io ?) will allow you to add more details as needed in terms of describing usage scenarios, details on examples, envisioned future extensions etc.
