### 2021-07-20 #3274 @vijaysm
@clavigne Can you add other workflow engines like Snakemake that are routinely used for distributed systems ? I am aware of [Fabsim](https://fabsim3.readthedocs.io/en/latest/index.html) which is a flexible workflow engine, though targeted at multiscale apps. It may be a bit more complex in terms of usage though. Another one is [Galaxy](https://training.galaxyproject.org/training-material/topics/admin/tutorials/connect-to-compute-cluster/tutorial.html). Your reference Molder et al., (2021) is for Snakemake, and I do not explicitly see any citations to "other" engines. It would be helpful to provide more context on what else is available with a similar feature set, and why funsies is different in the statement of need.
