### 2021-07-20 #3274 @danielskatz
There's also [Parsl](http://parsl-project.org) but I'm writing this more for your information than asking you to mention it, since I work on it.
