+++ paper/paper.bib
@@ -17,6 +17,14 @@
   url = {https://github.com/apache/airflow}
 }
 
+@misc{awesome_pipeline,
+  author = {Paolo Di Tommaso},
+  title = {A curated list of awesome pipeline toolkits inspired by Awesome Sysadmin},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/pditommaso/awesome-pipeline}
+}
 
 @misc{cadence,
   author = {Uber},
