### 2021-07-26 10d2932
added info about error handling in readme.md
