### 2021-07-26 #3274 @clavigne
Hi @vijaysm thanks again for the thorough review and I apologize for the delay in my reply. Hopefully I can provide satisfactory responses to your raised points, but please advise if you feel that this is not the case.

> Got it. I assumed the runtime error was intentional, but the log messages with trace information would confuse users, and lead them to worry that either the example is broken, or their build/installation is incompatible. So displaying this in a better, more user-friendly way would be better.

I changed the example to make the reporting (somewhat) more user-friendly. It now raises `StopRecursion` instead of `RuntimeError` and the behaviour is mentioned specifically in https://github.com/aspuru-guzik-group/funsies/blob/master/recipes/silly/README.md .

> You should also update README.md or appropriate documentation for funsies to describe how runtime errors are caught cleanly and redirected to the log file instead of workers quitting prematurely. This is an important feature for all real workflows and should be highlighted explicitly IMO.

That's an excellent point. I added this information to https://github.com/aspuru-guzik-group/funsies/blob/master/README.md . A more complete treatment will have to wait until the documentation is online (as described below) and I've finished designing the error handling. In particular, I'm still not entirely sure whether the current error handling is sufficiently ergonomic; it's the main part I want to stabilize at this time.

> Also, is there a separate documentation/example page for funsies ? I don't see one. There is a link to source docs and link to the recipes folder, but something like a user guide will add more value here as well. Bulk of what would go in a user document is already there in README.md, but separating this perhaps into its own space (github.io ?) will allow you to add more details as needed in terms of describing usage scenarios, details on examples, envisioned future extensions etc.

I agree, a separate documentation page is sorely needed. I haven't had the time to dive into it yet, and have been waiting on comments from other users about some of the ergonomics. I'll write a more complete documentation when I can free some time for it, hopefully before Fall.

> @clavigne Can you add other workflow engines like Snakemake that are routinely used for distributed systems ? I am aware of Fabsim which is a flexible workflow engine, though targeted at multiscale apps. It may be a bit more complex in terms of usage though. Another one is Galaxy. Your reference Molder et al., (2021) is for Snakemake, and I do not explicitly see any citations to "other" engines. It would be helpful to provide more context on what else is available with a similar feature set, and why funsies is different in the statement of need.

Yes! I cited Molder et al. because it contains the most thorough review and classification of those systems that I'm aware of. I've modified the paper to make this clearer,

> Scientific workflow management systems, many of which are compiled in [@awesome_pipeline] and systematically reviewed in [@molder_sustainable_2021], are easier to set up on high-performance computing clusters, but are tuned to the needs of specific disciplines, such as bioinformatics or machine learning.

Please advise if this solution is satisfactory. I didn't want to make the paper too long by reviewing what is (at this point) a very large number of alternatives (it's a bit of a crowded space).

Thanks again!
