### 2021-07-26 #3274 @clavigne
@danielskatz There's also Parsl but I'm writing this more for your information than asking you to mention it, since I work on it.

Thanks! I actually wasn't aware of Parsl when I embarked on this project, and by the time I did become aware of it, I already had done a fair bit of work. Parsl is quite similar to funsies (python-based, similar architecture, free licenses), albeit significantly more mature. There are probably interesting opportunities for collaboration and code sharing between the two projects (or at least for code I can borrow into funsies!)
