### 2021-07-26 #3274 @danielskatz
If you (very optionally) want to mention it, feel free to do so - we like citations, particularly of https://doi.org/10.1145/3307681.3325400 !

And if you want to discuss collaboration, we're happy to do that too - best might be to join the Parsl slack via http://parsl-project.org/support.html
