### 2021-08-11 16d0eff
reafactor get_db to get_redis
