+++ src/funsies/_constants.py
@@ -46,5 +46,3 @@ DAG_OPERATIONS = "funsies.dags.operations"
 DAG_STATUS = "funsies.dags.status.done"
 DAG_INDEX = "funsies.dags.index"
 
-# Max size of continuous data
-BLOCK_SIZE = 30 * 1024 * 1024  # 30 MB
