+++ src/funsies/_graph.py
@@ -15,10 +15,11 @@ from redis.client import Pipeline
 
 # module
 from . import _serdes
-from ._constants import _Data, ARTEFACTS, BLOCK_SIZE, Encoding, hash_t, join, OPERATIONS
+from ._constants import _Data, ARTEFACTS, Encoding, hash_t, join, OPERATIONS
 from ._funsies import Funsie
 from ._logging import logger
 from ._short_hash import hash_save
+from ._storage import StorageEngine
 from .config import Options
 from .errors import Error, ErrorKind, match, Result
 
@@ -160,7 +161,7 @@ def create_link(db: Redis[bytes], afrom: hash_t, ato: hash_t) -> None:
     assert is_artefact(db, ato)
     old = get_status(db, afrom)
     if old == ArtefactStatus.const:
-        logger.error("attempted to mark done a const artefact.")
+        logger.error("attempted to link a const artefact to something else.")
     else:
         set_status(db, afrom, ArtefactStatus.linked)
         key = join(ARTEFACTS, afrom, "links_to")
@@ -178,14 +179,9 @@ def resolve_link(db: Redis[bytes], address: hash_t) -> hash_t:
         return resolve_link(db, out)
 
 
-def _set_block_size(n: int) -> None:
-    """Change block size."""
-    global BLOCK_SIZE
-    BLOCK_SIZE = n
-
-
 def __get_data_loc(
-    store: Redis[bytes],
+    db: Redis[bytes],
+    store: StorageEngine,
     artefact: Artefact[Any],
     carry_error: Optional[hash_t] = None,
     do_resolve_link: bool = True,
@@ -194,10 +190,10 @@ def __get_data_loc(
     address = artefact.hash
 
     if do_resolve_link:
-        address = resolve_link(store, address)
+        address = resolve_link(db, address)
 
     # First check the status
-    stat = get_status(store, address)
+    stat = get_status(db, address)
 
     # if it's a link, we move over to the link
     if stat == ArtefactStatus.linked:
@@ -207,7 +203,7 @@ def __get_data_loc(
             source=carry_error,
         )
     elif stat == ArtefactStatus.error:
-        return Error.grab(store, address)
+        return Error.grab(db, address)
     elif stat <= ArtefactStatus.no_data:
         return Error(
             kind=ErrorKind.NotFound,
@@ -215,8 +211,8 @@ def __get_data_loc(
             source=carry_error,
         )
     else:
-        key = join(ARTEFACTS, address, "data")
-        if not store.exists(key):
+        key = store.get_key(address)
+        if key is None:
             return Error(
                 kind=ErrorKind.Mismatch,
                 details="expected data was not found",
@@ -226,41 +222,29 @@ def __get_data_loc(
             return key
 
 
-def get_bytes(
-    store: Redis[bytes],
-    source: Artefact[Any],
-    carry_error: Optional[hash_t] = None,
-    do_resolve_link: bool = True,
-) -> Result[bytes]:
-    """Retrieve bytes corresponding to an artefact."""
-    raw = match(
-        __get_data_loc(store, source, carry_error, do_resolve_link),
-        some=lambda x: b"".join(store.lrange(x, 0, -1)),
-        none=lambda x: x,
-    )
-    return raw
-
-
 # Save and load artefacts
 def get_data(
-    store: Redis[bytes],
+    db: Redis[bytes],
+    store: StorageEngine,
     source: Artefact[T],
     carry_error: Optional[hash_t] = None,
     do_resolve_link: bool = True,
 ) -> Result[T]:
     """Retrieve data corresponding to an artefact."""
-    raw = get_bytes(store, source, carry_error, do_resolve_link)
+    key = __get_data_loc(db, store, source, carry_error, do_resolve_link)
+    raw = store.take(key).read()
[…]

[…]
