+++ tests/test_artefacts.py
@@ -27,68 +27,78 @@ def test_artefact_add() -> None:
     assert c == b"bla bla"
 
 
-# def test_artefact_add_implicit() -> None:
-#     """Test adding implicit artefacts."""
-#     options()
-#     store = MockServer().runtime()
-#     art = _graph.variable_artefact(store, hash_t("1"), "file", Encoding.blob)
-#     out = _graph.get_data(store, art)
-#     assert isinstance(out, Error)
-#     assert out.kind == ErrorKind.NotFound
+def test_artefact_add_implicit() -> None:
+    """Test adding implicit artefacts."""
+    options()
+    server = MockServer()
+    db = server.new_connection()
+    store = RedisStorage(db)
+    art = _graph.variable_artefact(db, hash_t("1"), "file", Encoding.blob)
+    out = _graph.get_data(db, store, art)
+    assert isinstance(out, Error)
+    assert out.kind == ErrorKind.NotFound
 
 
-# def test_operation_pack() -> None:
-#     """Test packing and unpacking of operations."""
-#     opt = options()
-#     store = MockServer().runtime()
-#     a = _graph.constant_artefact(store, b"bla bla")
-#     b = _graph.constant_artefact(store, b"bla bla bla")
-#     fun = f.Funsie(
-#         how=f.FunsieHow.shell,
-#         what="cat infile",
-#         inp={"infile": Encoding.blob},
-#         out={"out": Encoding.json},
-#         extra={},
-#     )
-#     op = _graph.make_op(store, fun, {"infile": a}, opt)
-#     op2 = _graph.Operation.grab(store, op.hash)
-#     assert op == op2
+def test_operation_pack() -> None:
+    """Test packing and unpacking of operations."""
+    opt = options()
+    server = MockServer()
+    db = server.new_connection()
+    store = RedisStorage(db)
 
-#     with pytest.raises(AttributeError):
-#         op = _graph.make_op(store, fun, {}, opt)
+    a = _graph.constant_artefact(db, store, b"bla bla")
+    b = _graph.constant_artefact(db, store, b"bla bla bla")
+    fun = f.Funsie(
+        how=f.FunsieHow.shell,
+        what="cat infile",
+        inp={"infile": Encoding.blob},
+        out={"out": Encoding.json},
+        extra={},
+    )
+    op = _graph.make_op(db, fun, {"infile": a}, opt)
+    op2 = _graph.Operation.grab(db, op.hash)
+    assert op == op2
 
-#     with pytest.raises(AttributeError):
-#         # no inputs
-#         op = _graph.make_op(store, fun, {}, opt)
+    with pytest.raises(AttributeError):
+        op = _graph.make_op(db, fun, {}, opt)
 
-#     with pytest.raises(AttributeError):
-#         # too many inputs
-#         op = _graph.make_op(store, fun, {"infile": a, "infile2": b}, opt)
+    with pytest.raises(AttributeError):
+        # no inputs
+        op = _graph.make_op(db, fun, {}, opt)
 
-#     with pytest.raises(RuntimeError):
-#         op = _graph.Operation.grab(store, hash_t("b"))
+    with pytest.raises(AttributeError):
+        # too many inputs
+        op = _graph.make_op(db, fun, {"infile": a, "infile2": b}, opt)
+
+    with pytest.raises(RuntimeError):
+        op = _graph.Operation.grab(db, hash_t("b"))
 
 
-# def test_artefact_wrong_type() -> None:
-#     """Test storing non-bytes in implicit artefacts."""
-#     store = MockServer().runtime()
-#     art = _graph.variable_artefact(store, hash_t("1"), "file", Encoding.blob)
-#     _graph.set_data(
[…]
