+++ src/funsies/_storage.py
@@ -2,11 +2,8 @@
 from __future__ import annotations
 
 # std
-from enum import Enum
+from io import BytesIO
 from typing import Optional
-
-# external
-from redis import Redis
 
 # module
 from ._constants import hash_t
@@ -16,11 +13,20 @@ from .errors import Error, Result
 class StorageEngine:
     """Baseclass implementing the artefact storage protocol."""
 
-    def get_key(self, hash: hash_t) -> Optional[str]:
+    def get_key(self: StorageEngine, hash: hash_t) -> Optional[str]:
+        """Return the key for a given hash."""
         raise NotImplementedError("Baseclass used where derived class is required.")
 
-    def take(self, key: str) -> Result[bytes]:
+    def take(self: StorageEngine, key: str) -> Result[BytesIO]:
+        """Return a bytes stream for a given key.
+
+        Note: It is the caller's responsibility to .close() the stream.
+        """
         raise NotImplementedError("Baseclass used where derived class is required.")
 
-    def put(self, key: str, data: bytes) -> Optional[Error]:
+    def put(self: StorageEngine, key: str, data: BytesIO) -> Optional[Error]:
+        """Write stream data to a given key.
+
+        Note: this does not .close() the stream.
+        """
         raise NotImplementedError("Baseclass used where derived class is required.")
