+++ src/funsies/dynamic.py
@@ -9,7 +9,7 @@ from redis import Redis
 
 # module
 from ._constants import _Data, Encoding
-from ._context import get_options, get_redis
+from ._context import Connection, get_connection, get_options
 from ._graph import Artefact, constant_artefact, make_op
 from ._subdag import subdag_funsie
 from .config import Options
@@ -31,11 +31,11 @@ def sac(
     name: Optional[str] = None,
     strict: bool = True,
     opt: Optional[Options] = None,
-    connection: Optional[Redis[bytes]] = None,
+    connection: Connection = None,
 ) -> Artefact[T3]:
     """Perform a split/apply/combine dynamic DAG."""
     opt = get_options(opt)
-    db = get_redis(connection)
+    db, store = get_connection(connection)
 
     inputs: dict[str, Artefact] = {}
     # Parse input  -------------------------------------
@@ -59,9 +59,9 @@ def sac(
 
     def __sac(inpd: dict[str, Any]) -> dict[str, Artefact[T3]]:
         """Perform the map reduce."""
-        db = get_redis()
+        db, store = get_connection()
         args = [inpd[k] for k in arg_names]
-        split_data = [constant_artefact(db, d) for d in split_fun(*args)]
+        split_data = [constant_artefact(db, store, d) for d in split_fun(*args)]
         apply_data = [apply_fun(d) for d in split_data]
         combine_data = combine_fun(apply_data)
         return dict(out=combine_data)

[…]
