### 2021-08-12 0219753
need to fix storage sent to nodes
