+++ src/funsies/ui.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 # std
 import time
+import shutil
 from typing import Iterable, Mapping, Optional, overload, TypeVar, Union
 
 # external
@@ -17,17 +18,19 @@ except ImportError:
 
 # module
 from ._constants import _AnyPath, _Data, hash_t
-from ._context import get_options, get_redis
+from ._context import get_options, get_connection, Connection
 from ._dag import descendants, start_dag_execution
 from ._graph import (
     Artefact,
     constant_artefact,
     delete_artefact,
     get_data,
+    get_stream,
     get_status,
     make_op,
     Operation,
 )
+from ._storage import StorageEngine
 from ._logging import logger
 from ._run import is_it_cached
 from ._shell import shell_funsie, ShellOutput
@@ -42,18 +45,18 @@ _OUT_FILES = Optional[Iterable[_AnyPath]]
 T = TypeVar("T", bound=_Data)
 
 
-def _artefact(db: Redis[bytes], data: Union[T, Artefact[T]]) -> Artefact[T]:
+def _artefact(db: Redis[bytes],store:StorageEngine, data: Union[T, Artefact[T]]) -> Artefact[T]:
     if isinstance(data, Artefact):
         return data
     else:
-        return constant_artefact(db, server, data)
+        return constant_artefact(db, store, data)
 
 
 # --------------------------------------------------------------------------------
 # Dag execution
 def execute(
     *outputs: Union[Operation, Artefact, ShellOutput],
-    connection: Optional[Redis[bytes]] = None,
+        connection: Connection = None,
 ) -> None:
     """Trigger execution of a workflow to obtain a given output.
 
@@ -65,7 +68,7 @@ def execute(
             within a `Fun()` context.
     """
     # get redis
-    db = get_redis(connection)
+    db, _ = get_connection(connection)
 
     # run dag
     for el in outputs:
@@ -81,7 +84,7 @@ def shell(
     env: Optional[dict[str, str]] = None,
     strict: bool = True,
     opt: Optional[Options] = None,
-    connection: Optional[Redis[bytes]] = None,
+        connection: Connection=None,
 ) -> ShellOutput:
     """Add a shell command to the workflow.
 
@@ -140,7 +143,7 @@ def shell(
 
     """
     opt = get_options(opt)
-    db = get_redis(connection)
+    db, store = get_connection(connection)
 
     # Parse args --------------------------------------------
     cmds: list[str] = []
@@ -164,7 +167,7 @@ def shell(
                     + ' converted to json (and wrapped with "), \nyou NEED to pass it'
                     + " as bytes (by .encode()-ing it first)"
                 )
-            inputs[str(key)] = _artefact(db, val)
+            inputs[str(key)] = _artefact(db,store, val)
     else:
         raise TypeError(f"{inp} not a valid file input")
 
@@ -184,7 +187,7 @@ def shell(
 def put(
     value: T,
     *,
-    connection: Optional[Redis[bytes]] = None,
+        connection: Connection=None,
 ) -> Artefact[T]:
     """Save data to Redis and return an Artefact.
 
@@ -208,8 +211,8 @@ def put(
     Returns:
         An `types.Artefact` instance with status `const`.
     """
-    db = get_redis(connection)
-    return _artefact(db, value)
+    db,store = get_connection(connection)
+    return _artefact(db,store, value)
 
 
 def __log_error(where: hash_t, dat: Result[object]) -> None:
@@ -219,12 +222,12 @@ def __log_error(where: hash_t, dat: Result[object]) -> None:
 
 # fmt:off
 @overload
[…]
