### 2021-08-13 7b9caac
fixed tests with wait_for without timeouts
