### 2021-08-13 1f84799
typo in setup.py
