### 2021-08-13 444493d
added an easier initialization function
