### 2021-08-13 ce6fd08
finished automated funsies running
