### 2021-08-16 #3274 @vijaysm
@clavigne The article reads well and provides a clear description of the funsies package. Just one minor comment: consider adding a DOI for the citation Dean, J., Ghemawat, S. (2004). I could not find one for that version when I checked quickly. But the 2008 version of the paper has one. https://doi.org/10.1145/1327452.1327492

I should be able to wrap up the rest of my review in the next couple of days.
