### 2021-08-21 #3274 @clavigne
@diehlpk As I mentioned previously ( https://github.com/openjournals/joss-reviews/issues/3274#issuecomment-848041524 ) this specific article doesn't have DOI. The reviewer @vijaysm recommends instead citing the later review on MapReduce ( https://doi.org/10.1145/1327452.1327492 ) as opposed to the primary article ( https://www.usenix.org/legacy/events/osdi04/tech/dean.html ). This is an acceptable solution if citations must absolutely be associated with DOIs, but I'd rather cite the primary literature if possible which is why I did not make the change previously.

I'm willing to change the reference of course, but I feel like there is a bug in the JOSS software/process if references not accompanied with DOIs are rejected. It's not uncommon to cite DOI-less resources such as government reports or older papers (such as this one).
