### 2021-08-26 88a9c52
switch exception to print on bad start
