+++ tests/test_readmemd.py
@@ -21,8 +21,7 @@ def assert_in(thing: str, readme: str) -> None:
 init = r"""
 start-funsies \
     --no-pw \
-    --workers 2 \
-    --port 33333
+    --workers 2
 """
 
 # Script
@@ -66,10 +65,10 @@ def test_readmemd() -> None:
         readmemd = f.read()
 
     with tempfile.TemporaryDirectory() as dir:
-        # we use a different port otherwise this messes with github CI
-        # assert_in(init, readmemd)
+        assert_in(init, readmemd)
         start = subprocess.run(init.strip(), shell=True, cwd=dir)
         assert start.returncode == 0
+
         with open(join(dir, "hello-world.py"), "w") as f:
             f.write(script)
 
