### 2021-09-06 #3274 @vijaysm
Another small note is that I was not able to run your instructions above on Cori. For example, after installing funsies through `pip`, `start-funsies` was not callable. I thought perhaps this was installed with pip, but it's not available in my `$PATH`.

However, I was able to test the installation and runs on Cori at NERSC. The example workflow provided [here](https://github.com/aspuru-guzik-group/funsies/tree/3b4425457d51f9a90d3163f696bf8d8b51779d5d/recipes/slurm-conformers) worked cleanly with appropriate changes to the `slurm-submit.sh` file. @clavigne It would be good to add a link to the Slurm submit example in README.md or even provide a separate subsection there to talk about using funsies for workflows on LCF machines or clusters.

@diehlpk I've completed my review and I am satisfied with the software documentation and the accompanying paper. I was also able to test funsies installation and examples on a variety of different platforms without any issues.

I also started writing a new example to orchestrate one of my workflows in Climate data analysis, but did not make enough progress on that the last two weeks. That was another reason for the delay - but I think I'll continue that experiment separately and raise issues (if any) directly in the repository in the next couple of weeks.
