### 2021-09-07 #3274 @diehlpk
@clavigne Could you please check the blue names with the question mark at the end?

It seems that they are rendered as links, however, these are not links. Are these GitHub usernames? If so, the @ could be the problem.

```
We acknowledge testing by early users Cher-Tian Ser ((chertianser?)), Kjell Jorner96
((kjelljorner?)) and Gabriel dos Passos Gomes ((gabegomes?)). CL also thanks Chris97
Crebolder ((ccrebolder?)) f
```
