### 2021-09-07 #3274 @diehlpk
@clavigne

```
We intend to maintain funsies and of course welcome collaborations from contributors93
around the world.
```

I think it could be nice to refer to

https://github.com/aspuru-guzik-group/funsies/blob/master/CONTRIBUTING.md

in the sentence above.
