### 2021-09-07 #3274 @diehlpk
@clavigne I just got power back and will read your paper the next days, and we should be able to finalize your paper by the end of this week.
