### 2021-09-20 #3274 @diehlpk
@clavigne Thanks for the edits, please create

- [ ] A new version including all the requested changes by the reviewers
- [ ] Upload this version to Zenodo or FigShare. Note that the authors and the title have to match with the paper

and post the version and DOI here, so I can add them to the paper and recommend it for acceptance.
