### 2021-09-25 #3274 @clavigne
Hi @diehlpk , version 0.8.1 is the final version of the software for publication.

The Zenodo DOI is 10.5281/zenodo.5528107 (url: https://doi.org/10.5281/zenodo.5528107 )

Thank you,
Cyrille
