### 2021-09-27 af1c085
fixing cases and removing unneeded abstracts
