+++ paper/paper.bib
@@ -39,7 +39,6 @@
 	title = {Automatic discovery of chemical reactions using imposed activation},
     doi = {10.26434/chemrxiv.13008500.v2},
 	url = {/articles/preprint/Automatic_discovery_of_chemical_reactions_using_imposed_activation/13008500/1},
-	abstract = {Computational power and quantum chemical methods have improved immensely since computers were first applied to the study of reactivity, but the de novo prediction of chemical reactions has remained challenging. We show that complex reactions can be efficiently and autonomously predicted using chemical activation imposed by simple geometrical constraints. Our approach is demonstrated on realistic and challenging chemistry, such as a triple cyclization cascade involved in the total synthesis of a natural product and several oxidative addition reactions of complex drug-like molecules. Notably and in contrast with traditional hand-guided computational chemistry calculations, our method requires minimal human involvement and no prior knowledge of products or mechanisms. Imposed activation can be a transformational tool to screen for chemical reactivity and mechanisms as well as to study byproduct formation and decomposition.},
 	author = {Lavigne, Cyrille and dos Passos Gomes, Gabriel and Pollice, Robert and Aspuru-Guzik, Alan},
 	urldate = {2020-09-30},
 	date = {2020-09-29},
@@ -54,7 +53,6 @@
 	url = {https://aip.scitation.org/doi/full/10.1063/5.0006002},
 	doi = {10.1063/5.0006002},
 	shorttitle = {{PSI}4 1.4},
-	abstract = {{PSI}4 is a free and open-source ab initio electronic structure           program providing implementations of Hartree–Fock, density functional theory, many-body           perturbation theory, configuration interaction, density cumulant theory, symmetry-adapted           perturbation theory, and coupled-cluster theory. Most of the methods are quite efficient,           thanks to density fitting and multi-core parallelism. The program is a hybrid of C++ and           Python, and calculations may be run with very simple text files or using the Python {API},           facilitating post-processing and complex workflows; method developers also have access to           most of {PSI}4’s core functionalities via Python. Job specification may be passed           using The Molecular Sciences Software Institute ({MolSSI}) {QCSCHEMA} data format,           facilitating interoperability. A rewrite of our top-level computation driver, and           concomitant adoption of the {MolSSI} {QCARCHIVE} {INFRASTRUCTURE} project, makes the           latest version of {PSI}4 well suited to distributed computation of large numbers of           independent tasks. The project has fostered the development of independent software           components that may be reused in other quantum chemistry programs.},
 	pages = {184108},
 	number = {18},
 	journaltitle = {Journal of Chemical Physics},
@@ -65,12 +63,11 @@
 }
 
 @article{aquilante:2020,
-	title = {Modern quantum chemistry with [Open]Molcas},
+	title = {Modern quantum chemistry with {[Open]Molcas}},
 	volume = {152},
 	issn = {0021-9606},
 	url = {https://aip.scitation.org/doi/full/10.1063/5.0004835},
 	doi = {10.1063/5.0004835},
-	abstract = {{MOLCAS}/{OpenMolcas} is an ab initio electronic structure program providing a large set of computational methods from Hartree–Fock and density functional theory to various implementations of multiconfigurational theory. This article provides a comprehensive overview of the main features of the code, specifically reviewing the use of the code in previously reported chemical applications as well as more recent applications including the calculation of magnetic properties from optimized density matrix renormalization group wave functions.},
 	pages = {214117},
 	number = {21},
 	journaltitle = {Journal of Chemical Physics},
@@ -87,7 +84,6 @@
 	url = {https://aip.scitation.org/doi/full/10.1063/5.0007045},
 	doi = {10.1063/5.0007045},
 	shorttitle = {{CP}2K},
[…]
