### 2021-09-27 #3274 @danielskatz
@clavigne - I've suggested some small changes in https://github.com/aspuru-guzik-group/funsies/pull/5
(Note that removing the abstracts simply cleans up the XML file that we submit to Crossref - it has no effect on generated paper).

Let me know when you've merged these changes, and we can proceed to the final publishing step.

Also, and just FYI, this work is interesting to read in the context of the [Parsl project](http://parsl-project.org) to which I contribute.
