### 2021-10-02 #3274 @clavigne
@danielskatz @diehlpk Thanks a lot for your help! I merged in the changes. I think the paper is ready for publication.
