### 2021-10-02 #3274 @danielskatz
@clavigne - sorry, there's one more thing I missed initially and then just caught: https://github.com/aspuru-guzik-group/funsies/pull/6
