### 2021-10-02 #3274 @danielskatz
Note that at least, for me, the DOI is not yet resolving, so I'll keep this open until it does, the close the issue.
