### 2021-10-04 #3274 @danielskatz
The DOI now resolves, so we're done!!
