### 2025-08-08 #3274 @gflofst
Is this for a new release of the software? I remember reviewing this now that I dig in.
